requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
djangorestframework-simplejwt
pillow
django-phonenumber-field[phonenumberslite]
django-modeltranslation
django-ckeditor
django-filter
requests
django-auditlog


